+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Enhancements
 
+  * Documentation for dependencies under the `deps` directory is now be
+    included in the documentation search.
   * Inform about no documentation for the current search in minibuffer.
     It's not handy if workflow is cut apart when there is no
     documentation for a given search term.
@@ -18,6 +20,14 @@
   * Mark `alchemist-help-sexp-at-point` and
     `alchemist-help-module-sexp-at-point` as deprecated. The new function
     `alchemist-help-search-at-point` will adapt the functionality of both.
+
+### Bug fixes
+
+  * "No documentation found.." will be displayed in the minibuffer
+    when `TokenMissingError`, `SyntaxError` or `FunctionClauseError` apears in
+    the search output.
+  * Search term used to fail when dots and commas where at the end or beginning
+    of it. `__CALLER__.` / `List.to_integer/1,`
 
 ## v0.3.0